requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-environ==0.8.1
django-heroku==0.3.1
django-shortuuidfield==0.1.3
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
